Separate all the test case class definitions out
